suppressing warnings

+++ interface/apfsimWrapper.cc
@@ -35,19 +35,21 @@ namespace amsi
      * pMesh msh:      partitioned mesh.
      * apf::Field * u: displacement field. */
     double vol = 0.0;
-    /** Extract face entities of mesh that lie on face entities of geometric model */
+    // Extract face entities of mesh that lie on face entities of geometric model
     std::list<pEntity> ents;
     getClassifiedEnts(msh,pGF,2,std::back_inserter(ents));
-    /** for debugging purposes */
+    /*
+    // for debugging purposes
     AMSI_DEBUG (
       std::vector<int> tags;
       int ent_size = 0;
-      for (int ii=ent_size; ii<ents.size(); ii++)
+      for (size_t ii=ent_size; ii<ents.size(); ii++)
       {
         tags.push_back(GEN_tag(pGF));
         ent_size = ents.size();
       }
     )
+    */
     int surf_elem = 0;
     apf::Mesh * apfmsh = apf::getMesh(u);
     for(std::list<pEntity>::iterator ent = ents.begin(); ent != ents.end(); ++ent)
@@ -56,26 +58,28 @@ namespace amsi
       {
         pRegion mshRgn0 = F_region((pFace)(*ent),0);
         pRegion mshRgn1 = F_region((pFace)(*ent),1);
-	int normal_dir = 1;
-	int GRgn0_tag = mshRgn0 == NULL ? -2 : GEN_tag(R_whatIn(mshRgn0));
+        int normal_dir = 1;
+        int GRgn0_tag = mshRgn0 == NULL ? -2 : GEN_tag(R_whatIn(mshRgn0));
         int GRgn1_tag = mshRgn1 == NULL ? -2 : GEN_tag(R_whatIn(mshRgn1));
         if(GRgn0_tag == tag)
           normal_dir = 1;
         else if(GRgn1_tag == tag)
           normal_dir = -1;
-	else
-	{
-	  if(GRgn0_tag == -2 && GRgn1_tag != tag)
-	    normal_dir = 1;
-	  else if(GRgn1_tag == -2 && GRgn0_tag != tag)
-	    normal_dir = -1;
-	}
+        else
+        {
+          if(GRgn0_tag == -2 && GRgn1_tag != tag)
+            normal_dir = 1;
+          else if(GRgn1_tag == -2 && GRgn0_tag != tag)
+            normal_dir = -1;
+        }
+        /*
         AMSI_DEBUG (
           std::cout<<"surf_elem:"<<surf_elem<<" belongs to face: "<<tags[surf_elem]
           <<" Rgn "<<GRgn0_tag<<" inside,"
           <<" Rgn "<<GRgn1_tag<<" outside,"
           <<" normal direction "<<normal_dir<<std::endl;
         )
+        */
         vol += measureDisplacedFromSurf(apf::castEntity(*ent),u,normal_dir);
         surf_elem++;
       }
